fix: restructured email and sms notification logic
