+++ src/apps/backend/modules/notification/fcm-util.ts
@@ -14,7 +14,24 @@ import {
 export default class FcmUtil {
   private static firebaseApp: admin.app.App;
 
-  public static getFirebaseApp(): admin.app.App {
+  public static async sendPushNotification(
+    params: PushNotifcationParams
+  ): Promise<void> {
+    NotificationUtil.validatePushNotificationParams(params);
+    try {
+      const { fcmToken, body, title } = params;
+      this.getFirebaseApp();
+      const message = {
+        token: fcmToken,
+        notification: { title, body },
+      };
+      await admin.messaging().send(message);
+    } catch (err) {
+      throw new ServiceError(err as Error);
+    }
+  }
+
+  private static getFirebaseApp(): admin.app.App {
     if (!this.firebaseApp) {
       try {
         const serviceAccountPath = ConfigService.getValue<string>(
@@ -34,22 +51,5 @@ export default class FcmUtil {
       }
     }
     return this.firebaseApp;
-  }
-
-  public static async sendPushNotification(
-    params: PushNotifcationParams
-  ): Promise<void> {
-    NotificationUtil.validatePushNotificationParams(params);
-    try {
-      const { fcmToken, body, title } = params;
-      this.getFirebaseApp();
-      const message = {
-        token: fcmToken,
-        notification: { title, body },
-      };
-      await admin.messaging().send(message);
-    } catch (err) {
-      throw new ServiceError(err as Error);
-    }
   }
 }
