+++ src/apps/backend/modules/notification/sms-util.ts
@@ -1,0 +1,40 @@
+import { Twilio } from 'twilio';
+
+import { ConfigService } from '../config';
+
+import NotificationUtil from './internal/notification-util';
+import { SendSMSParams, ServiceError } from './types';
+
+export default class SmsUtil {
+  private static smsClient: Twilio;
+
+  public static async sendSMS(params: SendSMSParams): Promise<void> {
+    NotificationUtil.validateSms(params);
+
+    try {
+      const client = this.getSmsClient();
+
+      await client.messages.create({
+        to: NotificationUtil.phoneNumberToString(params.recipientPhone),
+        messagingServiceSid: ConfigService.getValue(
+          'twilio.messaging.messagingServiceSid'
+        ),
+        body: params.messageBody,
+      });
+    } catch (err) {
+      throw new ServiceError(err as Error);
+    }
+  }
+
+  private static getSmsClient(): Twilio {
+    if (this.smsClient) {
+      return this.smsClient;
+    }
+
+    this.smsClient = new Twilio(
+      ConfigService.getValue('twilio.verify.accountSid'),
+      ConfigService.getValue('twilio.verify.authToken')
+    );
+    return this.smsClient;
+  }
+}
